Ensure that the delayed startup observer is closed
Previously, the extension would keep on creating new observers for every
new window and not close them, leading to memory leaks.

This should also resolve the missing gBrowser errors being thrown by the
extension, which is caused by observers for closed windows that are
still getting notified because the observers were not closed.